Add initial implementation of Go CloudFlare ByPass round tripper
This commit includes the initial implementation of the Go CloudFlare ByPass round tripper, which is a small package designed to prevent triggering CloudFlare's "attention required" status for HTTP requests.

The round tripper achieves this by adding required/validated headers on requests and updating the client TLS configuration to pass CloudFlare validation.

Additionally, the commit includes a dependency on eddycjy/fake-useragent for setting a believable request user agent, which adds some variety for the long term.

The usage of the round tripper is documented in the README, and an important note has been added regarding the use of http.DefaultTransport.

Overall, this commit provides a functional implementation of the Go CloudFlare ByPass round tripper and sets a foundation for future development.